Started implementing geometry generation
Currently it draws the first lineDef. It looks like it's just solid color, but that's just because the UVs aren't being setup for each vertex yet. The material on the sub mesh does have the correct DOOM texture on it, though. :)

+++ Assets/Scripts/WADs/Data/Maps/SectorDef.cs
@@ -5,8 +5,8 @@ namespace DIY_DOOM.WADs.Data.Maps
 {
     public struct SectorDef
     {
-        public int FloorHeight;
-        public int CeilingHeight;
+        public float FloorHeight;
+        public float CeilingHeight;
 
         public string FloorTexture;
         public string CeilingTexture;
